Serialize outcome field as dict, not string.
The `df.to_json()` function returns serialized json string, which was
then serialized as string in the resulting JSON output.

This required consumers of the JSON to manually reparse the `outcome`
field as JSON for every output in order to use it.

This simple change still uses `to_json()`, to ensure that the outcome
field is JSON serialize able, but it parses it back into a python dict,
so that when the final JSON happens, we get the structured outcome field
in the JSON output, rather than a string.

Fixes #73